Initial proposed implementation for handling almost-good fishing holes. (#10936)
Change assessment of pond quality from binary to trinary - introducing a "suboptimal" pond, which may appear to a player to be valid, bit is not (because of underwater flowing water instead of water source blocks).
The fisher will give a notification (pasted below) about the quality of the pond, including location and guidance on how to repair it.